+++ resources/views/portal/mail/cancelled.blade.php
@@ -2,6 +2,6 @@
 
 {{ __('statamic-payments::portal.mail_cancelled_body', ['product' => $product, 'date' => $date, 'time' => $time]) }}
 
-{{ __('statamic-payments::portal.mail_cancelled_no_further') }}
+{{ $until ? __('statamic-payments::portal.mail_cancelled_until', ['date' => $until]) : __('statamic-payments::portal.mail_cancelled_no_further') }}
 
 {{ __('statamic-payments::portal.mail_cancelled_keep') }}